Fix:fixed issue with skipping computer turns
