+++ lib/turn.rb
@@ -9,46 +9,14 @@ class Turn
     return STDIN.gets.chomp
   end
 
-  def start_game
-    start_value = gets.chomp
-    if start_value == "p"
-      puts board.print_board
-      while true
-        turn = Turn.new(board)
-        input = turn.player_input.upcase
-        turn.player_piece(input)
-        # check if they have won
-        turn.computer_piece
-        # check if they have won
-        if turn.draw? == false
-          puts "It's a draw! ༼ つ ◕_◕ ༽つ Enter 'ruby ./lib/connect_four.rb' to play again!"
-          return
-        end
-      end
-    elsif start_value
-      puts "Please enter either p or q, please dont make us ask again..."
-      start_game
-
-    elsif start_value == "q"
-      puts "We wish you would enter p or q."
-        return
-    end
-  end
-
   def invalid_choice
-    # until  do
-    #   5.times
-      puts "We told you A-G. Try again!"
-      player_piece(gets.chomp.upcase)
-      return
-    # end
-    # puts "SERIOUSLY WE MEAN A-G"
+    puts "We told you A-G. Try again!"
+    player_piece(gets.chomp.upcase)
   end
 
   def another_choice
-      puts "Column is full. Please choose another column"
-      player_piece(gets.chomp.upcase)
-      return
+    puts "Column is full. Please choose another column"
+    player_piece(gets.chomp.upcase)
   end
 
   def draw?
@@ -270,9 +238,9 @@ class Turn
     end
   end
 
-  def computer_input
-    @possible_input = (%w(A B C D E F G)).sample
-  end
+  # def computer_input
+  #   @possible_input = (%w(A B C D E F G)).sample
+  # end
 
   def invalid_computer_choice
     # puts "redo"
@@ -280,31 +248,32 @@ class Turn
   end
 
   def computer_piece
-
+    @possible_input = (%w(A B C D E F G)).sample
+    computer_input = @possible_input
     if computer_input == "A"
       if @board.board[:row1][0] == "."
         @board.board[:row1][0] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row2][0] == "."
         @board.board[:row2][0] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row3][0] == "."
         @board.board[:row3][0] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row4][0] == "."
         @board.board[:row4][0] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row5][0] == "."
         @board.board[:row5][0] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row6][0] == "."
         @board.board[:row6][0] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       else
         invalid_computer_choice
@@ -313,27 +282,27 @@ class Turn
     elsif computer_input == "B"
       if @board.board[:row1][1] == "."
         @board.board[:row1][1] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row2][1] == "."
         @board.board[:row2][1] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row3][1] == "."
         @board.board[:row3][1] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row4][1] == "."
         @board.board[:row4][1] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row5][1] == "."
         @board.board[:row5][1] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row6][1] == "."
         @board.board[:row6][1] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       else
         invalid_computer_choice
@@ -343,27 +312,27 @@ class Turn
 
       if @board.board[:row1][2] == "."
         @board.board[:row1][2] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row2][2] == "."
         @board.board[:row2][2] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row3][2] == "."
         @board.board[:row3][2] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row4][2] == "."
         @board.board[:row4][2] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row5][2] == "."
         @board.board[:row5][2] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row6][2] == "."
         @board.board[:row6][2] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       else
         invalid_computer_choice
@@ -373,27 +342,27 @@ class Turn
 
       if @board.board[:row1][3] == "."
         @board.board[:row1][3] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row2][3] == "."
         @board.board[:row2][3] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row3][3] == "."
         @board.board[:row3][3] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row4][3] == "."
         @board.board[:row4][3] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row5][3] == "."
         @board.board[:row5][3] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row6][3] == "."
         @board.board[:row6][3] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       else
         invalid_computer_choice
@@ -403,27 +372,27 @@ class Turn
 
       if @board.board[:row1][4] == "."
         @board.board[:row1][4] = 'O'
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row2][4] == "."
         @board.board[:row2][4] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row3][4] == "."
         @board.board[:row3][4] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row4][4] == "."
         @board.board[:row4][4] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row5][4] == "."
         @board.board[:row5][4] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row6][4] == "."
         @board.board[:row6][4] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       else
         invalid_computer_choice
@@ -433,27 +402,27 @@ class Turn
 
       if @board.board[:row1][5] == "."
         @board.board[:row1][5] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row2][5] == "."
         @board.board[:row2][5] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row3][5] == "."
         @board.board[:row3][5] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row4][5] == "."
         @board.board[:row4][5] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row5][5] == "."
         @board.board[:row5][5] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row6][5] == "."
         @board.board[:row6][5] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       else
         invalid_computer_choice
@@ -463,27 +432,27 @@ class Turn
 
       if @board.board[:row1][6] == "."
         @board.board[:row1][6] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row2][6] == "."
         @board.board[:row2][6] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row3][6] == "."
         @board.board[:row3][6] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row4][6] == "."
         @board.board[:row4][6] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row5][6] == "."
         @board.board[:row5][6] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       elsif @board.board[:row6][6] == "."
         @board.board[:row6][6] = "O"
-        @board.print_board
+        @board.print_computer_board
 
       else
         invalid_computer_choice
